Invert KriegspielGame ownership
