+++ kriegspiel/wild16.py
@@ -4,7 +4,6 @@
 
 from __future__ import annotations
 
-from kriegspiel.berkeley import BerkeleyGame
 from kriegspiel.game import KriegspielGame
 from kriegspiel.move import KriegspielScoresheet as KSSS
 from kriegspiel.rulesets import RULESET_WILD16
@@ -18,10 +17,10 @@ class Wild16Game(KriegspielGame):
         super().__init__(ruleset=RULESET_WILD16)
 
     @classmethod
-    def _from_berkeley_game(cls, game):
+    def _from_kriegspiel_game(cls, game):
         """Rebuild a Wild16Game wrapper from a validated shared-engine instance."""
-        if not isinstance(game, BerkeleyGame):
-            raise TypeError("game must be a BerkeleyGame")
+        if not isinstance(game, KriegspielGame):
+            raise TypeError("game must be a KriegspielGame")
         if game.ruleset_id != RULESET_WILD16:
             raise ValueError("game must use the wild16 ruleset")
 
@@ -35,12 +34,14 @@ class Wild16Game(KriegspielGame):
         instance._blacks_scoresheet = KSSS.from_snapshot(game._blacks_scoresheet.snapshot())
         return instance
 
+    _from_berkeley_game = _from_kriegspiel_game
+
     @classmethod
     def from_snapshot(cls, snapshot):
         """Build a Wild16Game from a validated snapshot."""
-        return cls._from_berkeley_game(BerkeleyGame.from_snapshot(snapshot))
+        return cls._from_kriegspiel_game(KriegspielGame.from_snapshot(snapshot))
 
     @classmethod
     def load_game(cls, filename):
         """Load a Wild 16 game from disk."""
-        return cls._from_berkeley_game(load_game_from_json(filename))
+        return cls._from_kriegspiel_game(load_game_from_json(filename))
